3. Display "20 of 382 industry sectors" in small text below the
"Goods & Services" tile.

when we have multi-regional models => the number of sectors in
the model is different than the number of sectors that is
displayed. in this case, we currently take the sectors from the
result. but, when this is just used as a sector list, we want
to skip the calculation step, and thus may have to manually
aggregate the sectors then