display the ranking value
The heatmap result is based on a calculation result with a indicator * sector
matrix $R$ and a vector $t$ which contains a total indicator result that is used
to normalize $R$. In case of a multi-regional result, we first aggregate the
columns in $R$ (just by adding them up for sectors with the same sector code) to
transform it to a single-region result.

The first step is then to normalize the indicator results in $R$ with the
indicator totals in $t$:

$$ n_{i,j} = \frac{r_{i,j}}{t_{i}} $$

Then we calculate relative result shares by dividing each $n_{i,j}$ with the
maximum value of the normalized results for the indicator $i$:

$$ s_{i,j} = \frac{n_{i,j}}{\text{max}_{k = 1...n}n_{i,k}} $$

The result shares are basically just used to calculate the transparency of the
colors in the heatmap. The ranking value $rank_j$ of a sector $j$ for a set of
indicators $I$ is calculated as the Euclidian distance of the point
$p_j = (n_{i_1,j}, n_{i_2,j}, ..., n_{i_q,j}), i_k \in I$ and
$p_0 = (0, 0, ...,0)$:

$$ rank_j = \sqrt{\sum_{i_k \in I}{n_{i_k,j}^2}}$$