bind sort items
we may hold the sort indicator in local state

```ts
const [indicator, setIndicator] = React.useState<Indicator | null>(null);
```

we need to pass the indicators into the commodity list than, or store them
in the widget, as we do for the sectors; also we need to get the direct
indicator results from the API (and aggregate them in case of a multi-regional
model):

```ts
private indicators: Indicator[];
private matrixD: Matrix;

//...
private init() {
  this.indicators = await model.indicators();
  const matrixD = await model.matrix("D");
  // aggregate in case of multi-regional model
}

public getIndicatorResults(indicator: Indicator): number[] {
  return !indicator
    ? zeros(this.sectors.length)
    : matrixD.rowOf(indicator.index);
}
```

alternatively, we could just make the `matrixD` public.

then, generating the menu items. the sort by selection item should be only
added when the selection is not empty:

```tsx

// create the sort menu
const sortMenu= [];
if (Object.keys(selection).length > 0) {
  // only add the sort by selection
  // when the selection is not empty
  sortMenu.push(
    <MenuItem
      onClick={() => {
        setMenuElem(null); // close
        setSorter("selection");
        setIndicator(null);
      }}>
      Selected first
    </MenuItem>
  );
}

// alphabetical sorting
sortMenu.push(
  <MenuItem
    onClick={() => {
      setMenuElem(null); // close
      setSorter("alphabetical");
      setIndicator(null);
    }}>
    Alphabetical
  </MenuItem>
);

// sort items for the indicators
for (const indicator of indicators) {
  sortMenu.push(
    // ...
  );
}
```

it would be cool, if we could put this into a stateless function that gets
the state functions of the outer context as properties:

```tsx
const CommoditySortMenu = (props: {
  setMenuElem: (elem: null | HTMLElement) => void,
  setSortBy: (sorter: "alphabetical" | "selection" | "indicator") => void,
  setIndicator: (indicator: Indicator) => void
}) => {
  const items = //...;
  return (
    <>
      {items}
    </>
  );
};
```

also, we may want to define this type:

```ts
type CommoditySortByType =
  "alphabetical"
  | "selection"
  | "indicator";
```