the `iogrid` is getting complex
so it is better to move it into separate files. it now uses the simplified
indicator names so we need to pull our data from the test instance, not the
production one (as it does not yet support these fields.)

```bash
npm run download -- --endpoint https://smmtool.app.cloud.gov/api
```

then the `USEEIO` model needs to be renamed to `USEEIOv1.2`... need to upload this
also to the test server that hosts the widget pages.

the left pane is always initially empty...